Move code for saving the mock data, from handlers into the controller
